Create crop_adjust_overlay.py
This program takes an input of a thermal image and visible image of the same scene. Each image will have different resolution, aspect ratio and viewpoint of the scene. This code draws a rectangle on the visible image, allows the user to adjust the size of the rectangle by using either 'w' for bigger or 's' for smaller. The viewer can judge by eye if the visible image matches the thermal image. Once they match, the visible image resolution is adjusted and cropped, and the visible and thermal image are overlayed on top of one another.